+++ src/main/java/io/khasang/jrepetitor/controller/ItemController.java
@@ -2,7 +2,7 @@ package io.khasang.jrepetitor.controller;
 
 import io.khasang.jrepetitor.dto.ItemDTOInterface;
 import io.khasang.jrepetitor.entity.Item;
-import io.khasang.jrepetitor.model.ItemByQuestionIdRequestWrapper;
+import io.khasang.jrepetitor.model.wrappers.ItemByQuestionIdRequestWrapper;
 import io.khasang.jrepetitor.service.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
