+++ src/main/java/io/khasang/jrepetitor/controller/QuestionController.java
@@ -2,7 +2,7 @@ package io.khasang.jrepetitor.controller;
 
 import io.khasang.jrepetitor.dto.QuestionDTOInterface;
 import io.khasang.jrepetitor.entity.Question;
-import io.khasang.jrepetitor.model.QuestionByQuizIdRequestWrapper;
+import io.khasang.jrepetitor.model.wrappers.QuestionByQuizIdRequestWrapper;
 import io.khasang.jrepetitor.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
