+++ src/main/java/io/khasang/jrepetitor/controller/UserController.java
@@ -7,8 +7,8 @@ import io.khasang.jrepetitor.dto.impl.UserDTOImpl;
 import io.khasang.jrepetitor.entity.Profile;
 import io.khasang.jrepetitor.entity.User;
 import io.khasang.jrepetitor.service.UserService;
-import io.khasang.jrepetitor.model.CreationProfileStatusResponseWrapper;
-import io.khasang.jrepetitor.model.CreationUserStatusResponseWrapper;
+import io.khasang.jrepetitor.model.wrappers.CreationProfileStatusResponseWrapper;
+import io.khasang.jrepetitor.model.wrappers.CreationUserStatusResponseWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
