--- src/main/java/io/khasang/jrepetitor/model/CreationUserStatusResponseWrapper.java
@@ -1,41 +1,0 @@
-package io.khasang.jrepetitor.model;
-
-public class CreationUserStatusResponseWrapper {
-    private Boolean isLoginExist;
-    private Boolean isPhoneExist;
-    private Boolean isEmailExist;
-
-
-    public CreationUserStatusResponseWrapper() {
-    }
-
-    public Boolean isOk() {
-        return !(isLoginExist || isPhoneExist || isEmailExist);
-    }
-
-    public Boolean getLoginExist() {
-        return isLoginExist;
-    }
-
-    public void setLoginExist(Boolean loginExist) {
-        isLoginExist = loginExist;
-    }
-
-    public Boolean getPhoneExist() {
-        return isPhoneExist;
-    }
-
-    public void setPhoneExist(Boolean phoneExist) {
-        isPhoneExist = phoneExist;
-    }
-
-    public Boolean getEmailExist() {
-        return isEmailExist;
-    }
-
-    public void setEmailExist(Boolean emailExist) {
-        isEmailExist = emailExist;
-    }
-
-
-}
