--- src/main/java/io/khasang/jrepetitor/model/SelectedItemWrapper.java
@@ -1,13 +1,0 @@
-package io.khasang.jrepetitor.model;
-
-public class SelectedItemWrapper {
-    private Long selectedItemId;
-
-    public Long getSelectedItemId() {
-        return selectedItemId;
-    }
-
-    public void setSelectedItemId(Long selectedItemId) {
-        this.selectedItemId = selectedItemId;
-    }
-}
