+++ src/main/java/io/khasang/jrepetitor/service/impl/QuizTryServiceImpl.java
@@ -7,9 +7,9 @@ import io.khasang.jrepetitor.dao.UserDao;
 import io.khasang.jrepetitor.dto.QuizTryDTOInterface;
 import io.khasang.jrepetitor.dto.impl.QuizTryDTOImpl;
 import io.khasang.jrepetitor.entity.*;
-import io.khasang.jrepetitor.model.QuestionAnswerWrapper;
-import io.khasang.jrepetitor.model.SelectedItemWrapper;
-import io.khasang.jrepetitor.model.UserTryWrapper;
+import io.khasang.jrepetitor.model.wrappers.QuestionAnswerWrapper;
+import io.khasang.jrepetitor.model.wrappers.SelectedItemWrapper;
+import io.khasang.jrepetitor.model.wrappers.UserTryWrapper;
 import io.khasang.jrepetitor.service.QuizTryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
